Wall follow- 90 degree corners fixed
We made some changes to the code. Our problem last time was with 90 degree angles. We tested it out and the car was able to turn seems like a cmake error like last time. Speed automatically adjusts goes faster when there are no corners and slows down when it detects one. Corners need to be tuned if we want a faster time. Also we could not get dynamic reconfigure to work.